backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.7
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.0.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
reportlab==4.4.4
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
